Moved quiz components into folder and added basic styling
