+++ client/src/containers/QuizContainer.js
@@ -1,13 +1,14 @@
-import QuizList from "../components/QuizList"
-import QuizDetails from "../components/QuizDetails"
+import QuizList from "../components/QuizComponents/QuizList"
+import QuizDetails from "../components/QuizComponents/QuizDetails"
+import '../Quiz.css';
 
 const QuizContainer = ({organs, organToShow, showOrgan}) => {
 
     return (
-    <>
+    <div className="quizcontainer">
     <QuizList organs={organs} showOrgan={showOrgan}/>
     <QuizDetails organToShow={organToShow}/>
-    </>
+    </div>
  )
  
 }
